243: move requests under json module
